rename product api to marketplace api
