+++ api/v1/v1.go
@@ -11,7 +11,7 @@ import (
 	roleid "github.com/MyriadFlow/storefront_gateway/api/v1/roleId"
 	"github.com/MyriadFlow/storefront_gateway/api/v1/uploadtoipfs"
 	"github.com/MyriadFlow/storefront_gateway/api/v1/uploadtonfts"
-	"github.com/MyriadFlow/storefront_gateway/api/v1/products"
+	"github.com/MyriadFlow/storefront_gateway/api/v1/marketplace"
 	"github.com/gin-gonic/gin"
 )
 
@@ -29,6 +29,6 @@ func ApplyRoutes(r *gin.RouterGroup) {
 		uploadtonfts.ApplyRoutes(v1)
 		details.ApplyRoutes(v1)
 		healthcheck.ApplyRoutes(v1)
-		products.ApplyRoutes(v1)
+		marketplace.ApplyRoutes(v1)
 	}
 }
