+++ models/Marketplace.go
@@ -1,6 +1,6 @@
 package models
 
-type Product struct {
+type Marketplace struct {
 	ItemId              int   `gorm:"primary_key" json:"itemId"`
 	NFT_Contract_Address     string   `json:"nft-contract-address"`
 	TokenId            string   `json:"tokenId,omitempty"`
